Isaiah Fritch - Powerup Code - V2
--REQUESTS--
I need "lives" and "blocks" as a public static int so that I can interact with them, a "powers" list to store the powerups and code within the loop that checks when a block is destroyed.  When the block is destroyed, I need to check if the block holds a powerup and what the location of the center of the block is before it is removed, that way I can generate a powerup if needed.