Isaiah Fritch - Powerup Code - V10 05-04-23
Rewrote everything powerup related on the GameScreen to clean up some terrible code.  Now, the powerup is a lot cleaner.  There are a few small issues that I need to fix, but for now I have 'try and catches' around them.  The issues are very small, overall the code is seemingly done for power ups.  I just plan to clean it up and then move on to other things.